netlify-medvit/requirements.txt
torch==2.1.0+cpu
torchvision==0.16.0+cpu
Pillow==10.1.0
numpy==1.24.3
einops==0.7.0
timm==1.0.8
